fix: 确保训练标签为整数类型
- 在添加标签到数据集时使用 astype(int) 转换
- 修复 CrossEntropyLoss 要求标签为 Long 类型的问题
- 解决 'nll_loss_forward_reduce_cuda_kernel_2d_index' not implemented for 'Float' 错误

+++ lib/python/train_universal.py
@@ -241,14 +241,14 @@ class UniversalTrainer:
                 remove_columns=val_dataset.column_names
             )
 
-            # 添加标签
+            # 添加标签（确保是整数类型）
             train_dataset = train_dataset.add_column(
                 'labels',
-                train_df[label_column].tolist()
+                train_df[label_column].astype(int).tolist()
             )
             val_dataset = val_dataset.add_column(
                 'labels',
-                val_df[label_column].tolist()
+                val_df[label_column].astype(int).tolist()
             )
 
             # 加载模型
